+++ dbmodel/schemas/main/ud/base/ftrg/ud_gw_trg_edit_gully.sql
@@ -43,7 +43,6 @@ dx float;
 dy float;
 v_x float;
 v_y float;
-v_codeautofill boolean;
 v_srid integer;
 v_featurecat text;
 v_psector_vdefault integer;
@@ -69,7 +68,7 @@ v_gully_id text;
 v_childtable_name text;
 v_schemaname text;
 v_featureclass text;
-v_sys_code_autofill boolean;
+v_sys_code_autofill text;
 
 BEGIN
 
@@ -412,20 +411,23 @@ BEGIN
 			NEW.code := NULL;
 		END IF;
 
-				SELECT code_autofill, cat_feature.id, feature_class INTO v_code_autofill_bool, v_featurecat, v_featureclass
-		FROM cat_feature WHERE id=NEW.gully_type;
+		SELECT code_autofill, cat_feature.id, feature_class
+		INTO v_code_autofill_bool, v_featurecat, v_featureclass
+		FROM cat_feature WHERE id = NEW.gully_type;
 
 		IF NEW.code IS NULL THEN
 			NEW.code := gw_fct_generate_code('feature', NEW.gully_type, json_strip_nulls(row_to_json(NEW)::json));
+
+			IF NEW.code IS NULL AND v_code_autofill_bool THEN
+				NEW.code := NEW.gully_id;
+			END IF;
 		END IF;
 
-		IF (v_code_autofill_bool IS TRUE) AND NEW.code IS NULL THEN
-			NEW.code=NEW.gully_id;
-		END IF;
-
-		--Sys_code
-		IF v_sys_code_autofill IS TRUE THEN
-			NEW.sys_code=gen_random_uuid();
+		--Sys_code (uuid | code | none)
+		IF v_sys_code_autofill IN ('uuid', 'true') THEN
+			NEW.sys_code := gen_random_uuid();
+		ELSIF v_sys_code_autofill = 'code' THEN
+			NEW.sys_code := NEW.code;
 		END IF;
 
 		--Units