+++ dbmodel/schemas/main/ud/functions/ftrg/ud_gw_trg_edit_node.sql
@@ -86,7 +86,7 @@ v_cur_quadrant TEXT;
 v_new_lab_position public.geometry;
 v_dist_sign numeric;
 v_label_dist numeric;
-v_sys_code_autofill boolean;
+v_sys_code_autofill text;
 
 BEGIN
 
@@ -456,20 +456,23 @@ BEGIN
 			NEW.code := NULL;
 		END IF;
 
-				SELECT code_autofill, cat_feature.id INTO v_code_autofill_bool, v_featurecat
-		FROM cat_feature WHERE id=NEW.node_type;
+		SELECT code_autofill, cat_feature.id
+		INTO v_code_autofill_bool, v_featurecat
+		FROM cat_feature WHERE id = NEW.node_type;
 
 		IF NEW.code IS NULL THEN
 			NEW.code := gw_fct_generate_code('feature', NEW.node_type, json_strip_nulls(row_to_json(NEW)::json));
+
+			IF NEW.code IS NULL AND v_code_autofill_bool THEN
+				NEW.code := NEW.node_id;
+			END IF;
 		END IF;
 
-		IF (v_code_autofill_bool IS TRUE) AND NEW.code IS NULL THEN
-			NEW.code=NEW.node_id;
-		END IF;
-
-		--Sys_code
-		IF v_sys_code_autofill IS TRUE THEN
-			NEW.sys_code=gen_random_uuid();
+		--Sys_code (uuid | code | none)
+		IF v_sys_code_autofill IN ('uuid', 'true') THEN
+			NEW.sys_code := gen_random_uuid();
+		ELSIF v_sys_code_autofill = 'code' THEN
+			NEW.sys_code := NEW.code;
 		END IF;
 
 		-- Workcat_id